chore: first file restructuring
