+++ no-waste-fridge/src/screens/content/content_creation_button.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { TouchableOpacity, StyleSheet, Text } from "react-native";
-import colors from "../styles/colors";
+import colors from "../../styles/colors";
 
 export default function ContentCreationButton(props: any): React.JSX.Element {
     return (
