+++ no-waste-fridge/src/screens/content/content_grid.tsx
